Move private settings to a secret.xml resource
Instead of storing private keys in code, move them to a resource
file that is not checked into version control.

This is still not secure as this file can still be decompiled by anyone
with physical access to the device that is running the application,
but this is a first step to prevent us from storing private info
in version control.