Add token-based API authentication
Adds an ApiService that facilitates that all requests to the API are authenticated properly.

The builder returned by the ApiService ensures that all requests include the required API token
and if the token expires, automatically refreshes the token using basic auth based on the
username and password stored in secrets.xml.