Sync IdentificationEvents with back-end
Creates an abstract base class for all syncable models that contains the fields
needed for the syncing logic - a boolean on whether it has been synced and a
token field for storing the API token of the user who created the model.

The syncing is done in a separate service that checks for un-synced
IdentificationEvent models and marks them as synced if the POST is successful.