Fix member photo fetch to be more memory safe
After member fetch, query all members that still need the photo to be fetched
and store them in a queue and launch a thread that polls the queue and fetches
and stores the photo locally.

If the fetch photo thread is still running when the fetch member process runs
again, we launch an additional fetch photo thread which runs of the same queue
so we do not do duplicate work.