Better error messages for login failure
Stopped using an OkHttp Authenticator for adding basic auth header to
authentication token request because authenticator logic only gets called
after a 401 is returned resulting in an unnecessary first failing request.

Also improved the error messaging for login failures by specifying that
it is an invalid password if a 401 is returned, otherwise returns a generic
login failure message and reports the error to Rollbar.