Refactor TokenAuthenticator to TokenIntterceptor
First reason is that Authenticator will only call authenticate method after
a 401 response, so that it always causes an unnecessary API request without
authentication that will always fail.

Also update the interceptor logic to only use the API token stored in the
SharedPreferences only for GET requests because for all other API requests
we want to use the API token stored with the syncable model.

[#140553501]