More memory efficient member fetch
Previous logic was keeping a list of every member with photo thus causing memory
issues. This improved logic first deletes any members in the DB that do not exist
in the fetch results and then loops through the existing members and copies any
existing photo to the new member data (if the URL matches) and then cleans up the
memory before moving on to the next new member update.