Use ApplicationContext for DatabaseHelper
Passing getBaseContext from an activity to DatabaseHelper is dangerous because
it can reference the Activity which can cause memory leaks because the
DatabaseHelper singleton is used in services.

Instead pass an ApplicationContext because it is ok for the DatabaseHelper to be
tied to the application lifecycle and this will not reference the Activity.