Refactor enrollment fragments code
Create an abstract EnrollmentFragment class that contains all the boilerplate
logic shared between all enrollment fragments.

Adds the required enrollment fields to the Member model as well as some helper
methods for determining whether we should capture a fingerprint or photo of the
national ID of the member based on their age.

[#140651807]