Fix bug with null fingerprint guid or phone number in member sync
If a fingerprint guid or phone number is null, the request was still including
the key and using a blank string as the value which was failing the validation
in the back-end. Even passing a null RequestBody was still causing a blank
string to come through as the value, so updating the method to take different
signatures to ensure no value is being sent.

[#140829155]