Capture entire image during enrollment completion
Refactors the enrollment completion image capture logic to capture the entire
image instead of the thumbnail size image that was previously being captured.

In order to capture the entire image, a URI has to be passed to the camera
intent that points to a file that the intent can write to. This required
using a FileProvider to make the file accessible to an outside app (the
camera app).

Also includes logic to delete the locally stored image files once they have
been synced to the back-end.

There are still some potential timing issues with this approach, as we
do not immediately mark the member model as unsynced when we capture the
image (because we do not want the SyncService to sync just the photo), but
this does open up a window where the FetchService could overwrite the stored
URI if any member attributes get updated during that timespan.

[#140926789]