Allow syncing of arbitrary member fields
Updates the syncing logic to keep track of dirty fields by adding
a json array field to the Syncable model that keeps track of which
fields have been updated on the model.

When Member models are synced, it only includes the updated fields
in the PATCH request. For now, the Encounter and IdentificationEvent
models do not track dirty fields because the whole object is POSTed.

[#140912547]