Fix bug that was not allowing editing quantites for encounter items
The previous logic was adding an OnClickListener to encounter items with
service or lab billables to prevent the quantity from being edited. However
because the view containers (and thus views) are being re-used by the list view,
the holder with the listener was being re-used by encounter items with supplies,
drugs and vaccines and preventing the quantity from being edited.

Removed the listener that prevented the quantity change and instead enable/disable
the quantity field depending on whether the current billable type.

[#141200007]