Show detail view before encounter creation flow
When selecting a checked-in patient from the current patients list,
navigate to the patient detail view instead of new encounter view.

This is to provide more consistent UX so that everytime a member is
selected, it brings them to the same view so they can verify they selected
the correct member before taking an action.

This also prevents us from creating multiple open check-ins for a member.

[#140856977]