Fix bug in creating Member PATCH request
When we change photos locally, we do not change the local photo field
because the images are so large they have to be stored on the file system.
Therefore the previous logic would never mark the photo fields as dirty
and they would not be included in the PATCH request.

Instead mark the field dirty when the URL is changed to make sure they
get included.

[#141252815]