Add ability to enroll newborns
Adds a UI for enrolling newborns that requires entering a name,
card ID, birthdate, gender and photo.

Updates the Member model to store whether the member is new which
is needed to know whether a model represents a new member versus
an edited member. Also adds fields to specify a specific birthdate
as well as the corresponding birthdate accuracy.

Also starts taking advantage of the fact that a member is serializable
and can be passed between fragments instead of having to store it on the
activity - this allows us to pass a member between fragments without
having to persist it first.

[#141164141]