Pass serialized Member when navigating between fragmnets
Instead of storing a member ID on the activity, pass the entire serialized
member when navigating between fragments.

This avoids forcing the activity to maintain state and also removes having
to look up the member from the ID each time a fragment is created. It
also lets us pass un-saved members between fragments.

[#141164141]