Add UI for backdating an encounter
This allows the user to specify a date and time in the past to use as
the occurred_at timestamp when entering an encounter.

The main use case is for when members get treated over the weekend and
the data is manually recorded on paper until a clinic staff member that
can use the app is available that can enter the data into the app.

[#141169267]