Add backdated boolean field to Encounter
Adds a boolean field to Encounter to record whether the occurred_at
is a backdated valued.

Also updated the DatabaseHelper onUpgrade code to properly set the added
columns as not null with default values where applicable. This will not
impact devices that already have run the migrations, but will help in
making sure the current code in the migrations reflects best practices.

[#141169267]