Refactor Encounter logic to not hold state in Activity
Refactors the encounter logic to not require storing the current encounter
on the activity and instead passing the encounter object between the
encounter-related fragments.

This prevents bugs where the encounter data can get of sync if the user
backed out of the encounter flow and also addressed a bug where some
encounter attributes were persisting between adding different encounters.

[#141169267]