Fix bug with back button during encounter flow
If the user has attached an encounter form and hits the back button
from the receipt fragment, it was remembering the last attached
form but not displaying it so clicking finish would attach a second
copy of the form to the encounter.

Fixed by making sure a new EncounterForm is created each time
the EncounterFormFragment is loaded.