Report members that should be deleted instead of deleting them
The previous logic would prematurely delete any new members
(newborns) when new member data was fetched if the newborn had
not yet synced.

The new logic now checks whether the members not in the fetched set
are unsynced and skips them.

Additionally, instead of deleting members, it will notify Rollbar
that there is a member that should be able to be safely deleted so
that we can monitor this logic before updating it to actually
delete the member.

[#141995157]