Fix stale identification event data
When we added the ability to PATCH identification events, we
did not take into account that any saved identification events
on the device would become invalid because they were not being
properly marked as new when they were initially saved.

This is causing the syncing logic to attempt to PATCH
identification events that have not been created yet.

This commit adds a migration to fix the affected data on the phone
so that it can properly sync.

[#141995389]