Add Authenticator
The authenticator is a component for working with the Android accounts and
authentication frameworks. The goal of using the Android frameworks is to
leverage a standard way to authenticate users, be able to store tokens using
the Android AccountManager and because it is a necessary component to use
the Androind sync framework.

[#142990615]