Refactor authentication logic
This refactors our entire authentication/login logic. The approach is to leverage
the existing Android authentication framework, specifically the Authenticator
and AccountManager classes, to provide a more standard structure for handling
authentication.

The main change is separating out the authentication flow into its own Activity
that extends Android's AccountAuthenticatorActivity. This providers a cleaner
separation of business and authentication logic and also lets the login logic
be reusable. Having a separate activity also helps avoid subtle bugs when trying
to handle login/logged out state within the ClinicActivity - now the ClinicActivity
just checks that the logged in user has a token stored in the AccountManager
each time the activity starts and redirects to the AuthenticationActivity if it
is not present. This is also being done as a first-step towards integrating
with the Android Sync framework which requires an Authenticator to manage
authenticating before data is synced.

This integrates with Android's AccountManager by having each clinic user be
associated with an Android Account (the list of created accounts can be viewed
from the settings menu) and upon successful authentication with the back-end,
the returned authentication token is stored in the AccountManager and can be
retrieved when they are needed for making API requests.

Android's AccountManager does not handle remember what user (Account) is
currently logged in, so we have to manage that on our own which is accomplished
by storing the logged in user's username in the SharedPreferences upon login
and then clearing the username upon logout. We need to do this so the user does
not have to re-authentication (which requires an API request) each time they
interact with the app. This introduces a SessionManager class for encapsulating
the logging in, logging out and token fetching logic.

[#142990615]