Refactor sync logic to use JobService
Refactor the existing services to leverage Androids JobScheduler
which enables the Android OS to efficiently schedule when the
syncing logic runs by only running when there is a network
connection and batching jobs to preserve battery.

The one major change is that the minimum interval for jobs to be
scheduled is 15 minutes compared to the 1/10 minute intervals we
were using previously for fetch/sync respectively. I think this is
an acceptable change because Android updates are consistently enforcing
longer minimum intervals which means this should significantly help
with preserving battery - and becuase the app is optimized for offline,
this should not affect end-users in anyway.

[#142990671]