Store authentication token in ClinicActivity
Because fetching an authentication token from the SessionManager is an
asynchronous process, we do not want to have to handle that each time
we need a token to set a model as unsynced, so because we require a valid
authentication on initial sign-in to the ClinicActivity, we can store
that token and fetch it synchronously everytime we need it.

Because the FetchService is running in a background process, it is ok to
fetch the token asynchronously each time the service runs.

[#142990615]