Fix bug where fetch photo is being called on local URL
The call to fetchAndSetPhotoFromUrl within syncMember was occuring before the
URL was updated to URL from the response, so it was attempting to fetch using
the previous local URL and failing.

Updated the fetchAndSetPhotoFromUrl to not attempt to fetch local photos.

Refactored exception handling in some member methods so that they throw
the exceptions and let the caller choose how to handle.