Fix bug when enrolling newborn without a photo
Enrolling a newborn without taking a photo would cause a NPE to be thrown
when the member was synced.

This logic fixes the syncing logic to only attempt to include the photo in
the request if there is a photo URL present. It also updates the UI to disable
the complete enrollment button when enrolling a newborn until a photo is taken
because we always want the user to submit a photo during enrollment.

[#144886441]