First step in abstracting and simplifying syncing logic
Moves most of the syncing related logic into SyncableModel so that it is
consistent across all models. Simplifies the synced, dirty_fields and is_new
fields by just using dirty_fields and having synced and is_new be implicitly
derived from the dirty_fields - if dirty_fields is not empty, the model is
considered unsynced, and if the id field is marked as dirty, the model is
considered new.

Also consolidates the logic around how sync requests are formed by moving
all shared logic into SyncableModel and having abstract methods for the
model-specific syncing logic.

[#145314915]