Fix bugs in the syncing logic
Addresses three separate bugs in the syncing logic:

When formatting the patch request for the Member model, check if the photo
URL fields are dirty instead of the photo field because that is the field
that is changed when a photo is taken on the device.

Use the encounter ID instead of the encounter form ID when syncing the
encounter form.

In member, encounter and encounter form - make sure to set any fields on
the response to the same value as on the original model if they should not
be marked as dirty fields after the sync is performed.

[#145314915]