Throw exception when trying to delete file that does not exist
Update FileManager.deleteLocalPhoto to throw an exception if the specified
file does not exist on the device.

Update the code that calls this method to handle the exception but still
continue to mark the model as synced so that we do not continue to keep
trying to sync the model if the file does not exist.

[#145314915]