Do not generate IDs on syncable models
Setting generatedId to true was not doing what I expected, as it generating a
new ID for every model even after querying which was causing issues anytime
we wanted to patch an existing model.

Just setting id to true works for all models, because there is code in our
saveChanges logic to set an ID if one does not exist which will handle
generating IDs for new models.

[#145314915]