Refactor photo logic
Refactors our photo persisting logic to a cleaner design that helps us address issues
with our previous photo saving logic:

- cleans up after "orphaned" photos that are captured by the camera intent but never saved to the db
- deletes photos asynchronously in a service so that if the photo deletion fails, it will continue trying until it deletes the photo
- deletes the duplicate photo stored in the public gallery by the camera intent

The refactor accomplishes this by creating a separate model for handling the
necessary state for managing photos (tracking the url and when the photo is
synced and deleted). Also refactored some repeated logic in photo fragments
and moved it to the PhotoFragment. Registers a separate service responsible for
deleting the locally stored photos.

[#141928991]