Throw sync exception if we attempt to sync an empty request body map
There is a bug in our member syncing logic that can result in an empty request body map
being passed to the ApiService that results in an IllegalStateException which gets caught
outside the member syncing logic in the SyncService and prevents the service from syncing
the remaining un-synced members.

This change throws a more descriptive SyncException that includes information for us to
debug this edge case and the SyncException gets caught within the syncMembers loop so that
other un-synced members will continue to get synced.

[#150682740]