Fix timing bug with Billable syncing and fetching
This fix addresses a timing bug where a new billable is created as part of an
encounter but before it gets synced, new billables are fetched the newly created
billable gets deleted before syncing and causes a serialization error when
syncing the Encounter.

This fixes the bug by not deleting Billables where created_during_encounter is true.
These billables will still get deleted when necessary, because the next time they are
fetched, they will be returned in the response and be marked as created_during_encounter false.