Properly set up uploading test results and artifacts to Circle.
Right now we're only uploading test results and artifacts for our `app` module. We should be doing this for `device` and `domain` modules as well. I've commented out/disabled uploading test results for the `app` module for now as well since we don't have any tests for app yet, so the directory is never created and it's currently causing a "Directory not found" error on Circle.

[no tracker]