Set quantity EditTexts to white background:
The android:backgroundTint property that globally sets the EditText underline color to gray was also unintentionally setting the color of any EditText background drawable (e.g. view_outline_rectangle.xml) to gray. Since it seems that we'd want this gray underline anywhere we use an EditText, my first approach was to just overwrite the backgroundTint on the encounter item quantity EditText, but doing so completely overwrites the background (view_outline_rectangle.xml) itself.

[#158204636]