Move code from focusOut of EditText to focusIn on parent layout.
Currently, we have a bunch of code that hides the keyboard and makes views visible again whenever the user focuses out of an editText. This is what we want when a user focuses out of by pressing done or hiding the keyboard, but there's a third case where the user focuses out by clicking into *another* edit text (e.g. when they're toggling between editing different quantities) In this case, we don't want to hide the keyboard or make the hidden views visible just to have the onFocusIn of the new editText undo that work. This is what's happening currently and leading to some jumpy keyboard behavior.

[#158204636]