Fix bug: member edits don't show up on previous screens (such as CheckInMemberDetail) until they are reloaded (#306)
How to reproduce the issue: Search for a member, select the member, choose "Edit member", change their age, press back. Notice that the age is still the old age in the "CheckInMemberDetailFragment". Press back again, and the age is still the wrong age in the search results. Back out one more, search again, and the correct values will appear. 

This is especially noticeable for members missing fields. The "Needs to complete enrollment" warning will continue to appear until the screen is reloaded. 

What this PR does:
- changes are now reflected in the checkin member detail fragment correctly ("needs to complete enrollment" notification bar disapears after completing enrollment) and member info updates in that fragment.

What this PR is intentionally scoping out:
- Currently with these fixes, if a user edits a member name, and goes back to the *search member fragment*, the old info will still be there. I would like to scope this **updating the search member fragment** because it's a rare use case and it would involve a non-trivial change in how `SearchMemberViewModel` is implemented. 
  - It's a rare use case because the only reason someone would want to go back to search is if they clicked into the wrong member and want to select another member. It's super rare that they would edit a member, then use that search member fragment's results to select another member.