Fix behavior to auto-show keyboard on SearchMemberFragment (#309)
Fix logic to achieve UX of always showing keyboard upon loading the SearchMemberFragment.
Removed the auto-focus logic from the ToolbarSearch since it was not working, and makes
more sense for the Fragment to control that behavior. Also moved the initial ToolbarSearch
listeners to onCreateOptionsMenu instead of onPrepareOptionsMenu so it only gets called once.

In order to achieve the auto-show behavior, had to use the onPrepareOptionsMenu callback because
it did not work to hook into Fragment lifecycle methods since they play-out separately to the
the menu callback logic. Had to include a delay when posting the request to focus the View
because posting it instantly was preventing the keyboard from showing up which I am assuming
was due to other rendering happening within the ActionBar after that call.