Add pending claims option to menu.
I also introduce a new pattern where I decouple the ViewModel's ViewState into 2 separate chunks: a FormState to hold the state for the fields on the screen, and a MenuState to hold the state for the counts on the menu. This is to prevent unrelated parts of the page from triggering updates to each other - for example, before, the entire menu was getting invalidated and refreshing every time the user types in anything in any text field, which is very inefficient.

This commit also:
* reorganizes menu items to match latest designs
* updates menu item copy to match latest designs

[#160573283]