Optimize logic.
Use priceScheduleIssued flag on EncounterItem to filter down which price schedules to do db lookup on. We can actually write the logic to not have to do db lookups at all, but this is a bit more verbose. Because price schedule changes are rare, this solution is a trade-off between code simplicity and speed/efficiency. See https://github.com/Watsi/uhp-android-app/pull/465#discussion_r244251836 for full discussion.

[#162597724]