Require network before trying to sync for sandbox, demo, and staging build types. (#505)
Implementing the same thing from https://github.com/Watsi/coverage-enrollment/pull/323

Right now, only the `release` build type needs network for the services to run. However, that means if the build type is `sandbox`, or `demo`, or `staging`, we will still try to sync and fetch. This leads to a lot of spam in rollbar / slack notifications.

This PR makes it so that `release`, `staging`, `sandbox`, and `demo` build types need network before the sync / fetch services run. When testing locally (`debug` build with localhost:5000 on a laptop), the phone just doesn't need network since it can be connected straight to a laptop running backend.