[#164879447] Translate facilities receiving referrals (#531)
What this PR does:
- Translates the facility names into amharic / tigrinian

This is a slightly different approach than translating the other lists because this list is much longer and didn't want to change the type to an enum. Also "other" with custom text is an option here.
Listed questions throughout the comments.

After the release, I'd like to do a more holistic thinking around how we handle dropdowns across android, and backend because for example to add a referral reason to the dropdown, there would be a lot of files that we need to change across android and backend.