Clean up timezone handling (#579)
* Clean up timezone handling

This returns the injected Clock to use systemDefaultZone instead of UTC.
This is necessary because the timezone setting will affect how any logic
using that clock determines the date based on an instant of time. In order
to accurately show the user whether a timestamp corresponds to today for a
user, that timezone needs to match the current user_s expectation, which
should correspond to the system default.

This also updates the EthiopianDateHelper logic to only involve dates
so that it does not have to be conflated with timezone issues. It
instead relies on the caller to be responsible for passing in the
proper date value.