Fix bug: If admin panel user adds a non-health center billable type (like imaging, or bed_day or procedure), BillableApi conversion to Billable will fail. (#583)
If an admin panel user adds a billable (and a price) for a health center for a category like imaging, bed_day, or procedure, fetching will fail.

This PR keeps the android app flexible with which billable types it can fetch (even though there is no UI for it). This makes it so a misconfiguration of the system in the backend of price schedules does not cause fetching billables to fail.