Updated to the meso-secrets bucket (#633)
Created a new bucket on S3 called `meso-secrets` to replace `uhp-android-app` bucket.

This bucket contains the keys / secrets needed to deploy to play store for **both** clinic and enrollment apps. It felt a little weird naming the bucket `uhp-android-app` (which refers to clinic app) when secrets for both apps live there. Also wanted to be super clear this bucket holds secrets and not other things like member photos.

Also prefixes each `variables.gradle` with the app and flavor.