Remove flavor specific versions of AuthenticationActivity and main.xml (#645)
The only difference between the two AuthenticationActivity is one of them can "switch language" (Ethiopia) and the other one doesn't have (Uganda). For Demo, we'd probably want to keep the switch language functionality. However, in the code, we duplicate `AuthenticationActivity` and when we add demo, we would need to make a copy of `AuthenticationActivity` again. 

What this PR does:
- Combines the `AuthenticationActivity` and `main.xml` into the main flavor, so that when we add a demo flavor, we wouldn't need to duplicate this file.
- Adds a feature flag "ENABLE_SWITCH_LANGAUGE" as a temporary thing flag to differentiate whether to allow the user to switch language.

In a future PR:
- Merging `ClinicActivity` into the same fragment 
- Ability to switch language is probably something we'd want to support in all implementations of Meso, so the next step is moving that into a settings menu / dialog where the user can choose from a list of supported language.