Remove demo and training logic / build type (#647)
Getting ready so we don't confuse the demo build **type** to the demo build **flavour**.

This is also code that isn't used anymore (we used to have a banner at the top that says "training only" for Sr. Juliet to train end users but that is not a real build type anymore and we've have since moved to 
using sandbox for this purpose.