Set usesCleartextTraffic to false so debug mode works for android 9 pie OS
For android devices with API level 28 or higher, this value defaults to "false". As a result, when using the app in debug mode (i.e. things that hit http://localhost:5000) on android and app http requests will fail. This sets the clear text flag to true so debug mode works for android 9 pie OS.