Set usesCleartextTraffic to false when hitting localhost (#4)
Ran into an issue where I wasn't able to get debug mode to hit a server hosted on http://localhost:5000.

For android devices with API level 28 or higher, `userClearTextTraffic` defaults to "false". As a result, when using the app in debug mode (i.e. things that hit http://localhost:5000) on android and app http requests will fail. This sets the clear text flag to true so if we're hitting a localhost on a android 8.0+ device, it will work.

More info in issue here:
- https://stackoverflow.com/questions/45940861/android-8-cleartext-http-traffic-not-permitted